# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SI GAPLEK - Sistem Gudang dan Pengelolaan Logistik Kantor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock-in:
    get:
      operationId: listStockIn
      tags: [stock-in]
      summary: List stock-in transactions
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockIn"
    post:
      operationId: createStockIn
      tags: [stock-in]
      summary: Create stock-in transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockInBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockIn"
  /stock-in/{id}:
    get:
      operationId: getStockIn
      tags: [stock-in]
      summary: Get a stock-in transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockInDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stock-in/{id}/finalize:
    post:
      operationId: finalizeStockIn
      tags: [stock-in]
      summary: Finalize stock-in (updates stock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockIn"
        "400":
          description: Already finalized or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    StockIn:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        totalItems:
          type: integer
        transactionDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - referenceNo
        - status
        - totalItems
        - transactionDate
        - createdAt
    StockInDetail:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        transactionDate:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockTransactionItem"
      required:
        - id
        - referenceNo
        - status
        - transactionDate
        - createdAt
        - items
    CreateStockInBody:
      type: object
      properties:
        supplierId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        transactionDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockTransactionItemInput"
      required:
        - transactionDate
        - items
...
    StockTransactionItem:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        unitName:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - itemId
        - itemCode
        - itemName
        - quantity
        - unitPrice
    StockTransactionItemInput:
      type: object
      properties:
        itemId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        locationId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - itemId
        - quantity
        - unitPrice
...